requirements.txt
# Core API
fastapi==0.109.1
uvicorn==0.23.0
gunicorn==21.2.0
pydantic==2.2.1
pydantic_core==2.6.1
annotated-types==0.7.0
anyio==4.4.0
sniffio==1.3.1
h11==0.14.0
click==8.1.7
typing_extensions==4.12.2
aiofiles==23.2.1

# ML & Data
pandas==2.0.3
numpy==1.25.2
scikit-learn==1.3.0
scipy==1.14.0
joblib==1.4.2
threadpoolctl==3.5.0

# Fuzzy matching
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.1
Levenshtein==0.25.1
rapidfuzz==3.9.6

# HTTP & Config
requests==2.31.0
python-dotenv==1.0.0
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
urllib3==2.2.2

# Dates
python-dateutil==2.9.0.post0
pytz==2024.1
tzdata==2024.1
six==1.16.0
packaging==24.1
